Always mark full lines with wrong streak sizes as completely unsatisfied
Streak checks happen in normal and in reverse order. On full lines this
means that all streaks are matched twice against the solution, yielding
markers that may feel counter-intuitive.

In particular, a solution that consists of two identical sub-patterns is
marked as completely satisfied on a full line that contains the sub-pattern
only once. A board may even seem completely solved, while it is not.

To prevent this, extend the preliminary checks to completely reject full
lines with the wrong number of streaks.

BUG: 435211